+++ src/components/30-organisms/footer/react/demo.jsx
@@ -6,7 +6,6 @@ import changelog from '../CHANGELOG.md';
 import AXAFooterReact from './Footer';
 
 const setNewlyClickedLink = link => {
-  console.log('clickediclick');
   document.getElementById(
     'clicked-link'
   ).innerText = `Last clicked link: ${link}`;
